+++ packages/compiler/src/render3/view/util.ts
@@ -131,3 +131,27 @@ export class DefinitionMap {
 
   toLiteralMap(): o.LiteralMapExpr { return o.literalMap(this.values); }
 }
+
+/**
+ * Extract a map of properties to values for a given element or template node, which can be used
+ * by the directive matching machinery.
+ *
+ * @param elOrTpl the element or template in question
+ * @return an object set up for directive matching. For attributes on the element/template, this
+ * object maps a property name to its (static) value. For any bindings, this map simply maps the
+ * property name to an empty string.
+ */
+export function getAttrsForDirectiveMatching(elOrTpl: t.Element | t.Template):
+    {[name: string]: string} {
+  const attributesMap: {[name: string]: string} = {};
+
+  elOrTpl.attributes.forEach(a => {
+    if (!isI18NAttribute(a.name)) {
+      attributesMap[a.name] = a.value;
+    }
+  });
+  elOrTpl.inputs.forEach(i => { attributesMap[i.name] = ''; });
+  elOrTpl.outputs.forEach(o => { attributesMap[o.name] = ''; });
+
+  return attributesMap;
+}
